fix(chsql,chplan,logql): move AttrStrategy to chplan, fix arch-lint violation
internal/logql is architecturally forbidden from depending on
internal/chsql, but Lang.EmitAttrStrategies() returned chsql.AttrStrategies,
so logql/lang.go imported chsql directly. Moved the type definitions to
internal/chplan (already a dependency of both logql and chsql) and made
chsql re-export them as type/const aliases, so every existing
chsql.AttrStrategy* reference keeps compiling unchanged.
